Re-implement a lookup algorithm to get better hits.
The original CTakes permutation algorithm leaves the first word of the
permutation intact which means if your term is "a b c" but the user
typed "b a c" the term isn't gonna turn up.

Reimplemented the algorithm so that it also moves the first word.
Unfortunately the design of ctakes required that we bring in an awful
lot of extra code for a relatively small change.